requirements.txt
azure-functions
pymongo
pandas
requests
sendgrid
sendgrid-mail